==> 01_schola-styled-word.Rmd <==
---
title: "Schola-styled Word document"
subtitle: "A great one, too"
author:
- tým Schola Empirica
- "kontakt: Jaromír Mazák, mazak@scholaempirica.org"
date: '`r format(Sys.time(), "%d. %B %Y")`'
output: 
  reschola::schola_word:
    fig_caption: yes
    toc: no
    number_sections: false
abstract: Lorem ipsum abstract
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(include = T) # by default, chunk code and output is hidden
options(scipen = 10) # force R to show full numbers, not scientific notation

library(reschola)
library(tidyverse)
library(scales)
source("shared.R") # uncomment if working in a reschola project with a shared.R script
set_reschola_ggplot_fonts() # make ggplot2 use Roboto fonts without you having to set it
library(here)
library(corrr)
library(lme4)
library(lmerTest)
library(broom.mixed)
library(flextable)
library(ggeffects)
library(conflicted)
conflict_prefer("lmer", "lmerTest")
conflict_prefer("filter", "dplyr")
```

```{r}
dfm <- read_rds(here("data-processed/waves_1-7_cleaned-plus-IRT.rds"))
```


```{r}
dfm %>%
  drop_na(gender_girl) %>%
  mutate(wave_ind = as.factor(wave_ind)) %>%
  select(id_pupil, gender_girl, is_intervention2, wave_ind, emo:pro) %>%
  pivot_longer(emo:pro) %>%
  ggplot(aes(value, col = wave_ind)) +
  geom_density() +
  facet_grid(gender_girl ~ name) +
  theme_schola("x", multiplot = TRUE, legend.position = "bottom") +
  guides(colour = guide_legend(nrow = 1)) +
  labs(title = "Latent trait estimates", subtitle = "by gender, SDQ dimension and measurement number")
```

```{r}
dfm %>% filter(wave_ind == 1) %>% 
  drop_na(gender_girl) %>%
  pivot_longer(emo:pro) %>%
  ggplot(aes(age, value, col = gender_girl)) +
  geom_jitter(alpha = .1) +
  geom_smooth() +
  facet_grid(is_intervention2 ~ name) +
  theme_schola("scatter") +
  labs(title = "Věk a dimenze SDQ", subtitle = "první měření")
```

```{r}
dfm %>%
  select(id_pupil, gender_girl, is_intervention2, wave_ind, emo:pro) %>%
  pivot_longer(emo:pro) %>%
  ggplot(aes(x = wave_ind, value, col = is_intervention2)) +
  geom_line(aes(group = id_pupil), alpha = .06) +
  geom_smooth(method = "loess", se = FALSE, size = 1) + # fit may be unreliable!
  facet_grid(is_intervention2 ~ name) +
  theme_schola("x", multiplot = TRUE, legend.position = "bottom") +
  labs(title = "Jednotlivé trajektorie a průměrný trend", col = "skupina")
```

*Test* tabulek -- time dependency

```{r}
time_dependency <- function(df, dim, method = "pearson") {
  prefix <- dim
  dim <- enquo(dim)
  
  df %>% pivot_wider(id_cols = c(id_pupil), names_from = wave_ind, values_from = !!dim, names_prefix = paste0(str_extract(prefix, "[:alpha:]+"), "_")) %>% select(-id_pupil) %>% corrr::correlate(method = method) %>% corrr::shave(upper = FALSE) %>% corrr::fashion()
}

time_dependency(dfm, "pro", "spear") %>% flextable()

# purr and flextable -- to be solved
map(dfm %>% select(emo:pro) %>% colnames, ~ time_dependency(dfm, ., method = "spearman") %>% regulartable)
```


```{r}
# first measure to 0 - meaningful intercept - mean outcome for 1st timepoint
df_long_zero <-
  dfm %>% mutate(wave_ind = as.numeric(wave_ind) - 1)
```

## Null model

```{r}
null_model <- lmer(emo ~ (1 | id_pupil), data = df_long_zero)
summary(null_model)
```

## random intercept + slope

pupils are not yet nested within classes, model fails to converge

TODO: check id_class

```{r}
ris_model <- lmer(pro ~ wave_ind * is_intervention2 + age_fst_measur + gender_girl + (wave_ind | id_pupil), data = df_long_zero)
summary(ris_model)

df_long_zero %>% drop_na(ris_model@frame %>% colnames) %>%
  mutate(model_residuals = resid(ris_model), pred_vals = predict(ris_model)) %>%
ggplot(aes(pred_vals, model_residuals)) +
  geom_point(alpha = .15) +
  geom_vline(xintercept = 0, col = "red") +
  geom_hline(yintercept = 0, col = "red") +
  theme_schola("scatter") +
  labs(title = "Reziduály fitovaného modelu")

```


```{r}
resid(ris_model) %>%
  as_tibble %>%
  mutate(id = 1:n()) %>%
  ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  theme_schola("scatter") +
  labs(title = "Q-Q plot reziduálů")

# residuals densities
df_long_zero %>% drop_na(ris_model@frame %>% colnames) %>%
  mutate(model_residuals = residuals(ris_model)) %>%
ggplot(aes(x = model_residuals)) +
  geom_density(aes(color = is_intervention2), size = 1.25) +
  xlab("Residuals") +
  theme_schola("x") +
  labs(title = "Hustoty reziduálů dle skupiny")

# random effects of id_pupils to long format
ranef(ris_model)$id_pupil %>%
  mutate(id = 1:n()) %>%
  pivot_longer(-id, "variable", "value") %>%
ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  facet_wrap(~ variable, scales = 'free_y') +
  theme_schola("scatter") +
  labs(title = "Q-Q ploty random efektů")

```

## Vizualizace modelu -- průměry 

```{r}
df_long_zero %>%
  drop_na(ris_model@frame %>% colnames) %>%
  mutate(pred_val = predict(ris_model, re.form = NA)) %>% # fixed eff only
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = pro,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data - fixed effects")

df_long_zero %>%
  drop_na(ris_model@frame %>% colnames) %>%
  mutate(pred_val = predict(ris_model)) %>%
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = pro,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data")
```

## random effect for sample of 50 pupils

```{r}
#rand effects
tidy(ris_model, effects = "ran_vals", conf.int = TRUE) %>% sample_n(50) %>% 
ggplot(aes(level, estimate)) +
  geom_pointrange(aes(
    ymin = estimate - 1.96 * std.error,
    ymax = estimate + 1.96 * std.error
  )) + coord_flip() +
  theme_schola("x") +
  labs(title = "50 náhodně vybraných interceptů")
```

```{r}
ris_model %>% ggpredict()
```

{marginal plots draft

```{r}
# CI, fixed eff.
ris_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl")) %>%
  plot + theme_schola("scatter")

ris_model %>%
  ggpredict(terms = c("is_intervention2", "gender_girl", "age_fst_measur [meansd]")) %>% plot + theme_schola("scatter")

# one of the most interesting plots:
ris_model %>%
  ggpredict(terms = c("age_fst_measur", "is_intervention2", "wave_ind")) %>%
  plot + theme_schola("scatter")

# predictiton interval (random effects)
ris_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl"), type = "re") %>%
  plot + theme_schola("scatter")
```







# emo

```{r}
emo_model <- lmer(emo ~ wave_ind * is_intervention2 + age_fst_measur + gender_girl + (wave_ind | id_pupil), data = df_long_zero)
summary(emo_model)

df_long_zero %>% drop_na(emo_model@frame %>% colnames) %>%
  mutate(model_residuals = resid(emo_model), pred_vals = predict(emo_model)) %>%
ggplot(aes(pred_vals, model_residuals)) +
  geom_point(alpha = .15) +
  geom_vline(xintercept = 0, col = "red") +
  geom_hline(yintercept = 0, col = "red") +
  theme_schola("scatter") +
  labs(title = "Reziduály fitovaného modelu")

```


```{r}
resid(emo_model) %>%
  as_tibble %>%
  mutate(id = 1:n()) %>%
  ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  theme_schola("scatter") +
  labs(title = "Q-Q plot reziduálů")

# residuals densities
df_long_zero %>% drop_na(emo_model@frame %>% colnames) %>%
  mutate(model_residuals = residuals(emo_model)) %>%
ggplot(aes(x = model_residuals)) +
  geom_density(aes(color = is_intervention2), size = 1.25) +
  xlab("Residuals") +
  theme_schola("x") +
  labs(title = "Hustoty reziduálů dle skupiny")

# random effects of id_pupils to long format
ranef(emo_model)$id_pupil %>%
  mutate(id = 1:n()) %>%
  pivot_longer(-id, "variable", "value") %>%
ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  facet_wrap(~ variable, scales = 'free_y') +
  theme_schola("scatter") +
  labs(title = "Q-Q ploty random efektů")

```

## Vizualizace modelu -- průměry 

```{r}
df_long_zero %>%
  drop_na(emo_model@frame %>% colnames) %>%
  mutate(pred_val = predict(emo_model, re.form = NA)) %>% # fixed eff only
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = emo,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data - fixed effects")

df_long_zero %>%
  drop_na(emo_model@frame %>% colnames) %>%
  mutate(pred_val = predict(emo_model)) %>%
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = emo,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data")
```

## random effect for sample of 50 pupils

```{r}
#rand effects
tidy(emo_model, effects = "ran_vals", conf.int = TRUE) %>% sample_n(50) %>% 
ggplot(aes(level, estimate)) +
  geom_pointrange(aes(
    ymin = estimate - 1.96 * std.error,
    ymax = estimate + 1.96 * std.error
  )) + coord_flip() +
  theme_schola("x") +
  labs(title = "50 náhodně vybraných interceptů")
```

```{r}
emo_model %>% ggpredict()
```

marginal plots draft

```{r}
# CI, fixed eff.
emo_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl")) %>%
  plot + theme_schola("scatter")

emo_model %>%
  ggpredict(terms = c("is_intervention2", "gender_girl", "age_fst_measur [meansd]")) %>% plot + theme_schola("scatter")

# one of the most interesting plots:
emo_model %>%
  ggpredict(terms = c("age_fst_measur", "is_intervention2", "wave_ind")) %>%
  plot + theme_schola("scatter")

# predictiton interval (random effects)
emo_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl"), type = "re") %>%
  plot + theme_schola("scatter")
```

# con


```{r}
con_model <- lmer(con~ wave_ind * is_intervention2 + age_fst_measur + gender_girl + (wave_ind | id_pupil), data = df_long_zero)
summary(con_model)

df_long_zero %>% drop_na(con_model@frame %>% colnames) %>%
  mutate(model_residuals = resid(con_model), pred_vals = predict(con_model)) %>%
ggplot(aes(pred_vals, model_residuals)) +
  geom_point(alpha = .15) +
  geom_vline(xintercept = 0, col = "red") +
  geom_hline(yintercept = 0, col = "red") +
  theme_schola("scatter") +
  labs(title = "Reziduály fitovaného modelu")

```


```{r}
resid(con_model) %>%
  as_tibble %>%
  mutate(id = 1:n()) %>%
  ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  theme_schola("scatter") +
  labs(title = "Q-Q plot reziduálů")

# residuals densities
df_long_zero %>% drop_na(con_model@frame %>% colnames) %>%
  mutate(model_residuals = residuals(con_model)) %>%
ggplot(aes(x = model_residuals)) +
  geom_density(aes(color = is_intervention2), size = 1.25) +
  xlab("Residuals") +
  theme_schola("x") +
  labs(title = "Hustoty reziduálů dle skupiny")

# random effects of id_pupils to long format
ranef(con_model)$id_pupil %>%
  mutate(id = 1:n()) %>%
  pivot_longer(-id, "variable", "value") %>%
ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  facet_wrap(~ variable, scales = 'free_y') +
  theme_schola("scatter") +
  labs(title = "Q-Q ploty random efektů")

```

## Vizualizace modelu -- průměry 

```{r}
df_long_zero %>%
  drop_na(con_model@frame %>% colnames) %>%
  mutate(pred_val = predict(con_model, re.form = NA)) %>% # fixed eff only
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = con,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data - fixed effects")

df_long_zero %>%
  drop_na(con_model@frame %>% colnames) %>%
  mutate(pred_val = predict(con_model)) %>%
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = con,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data")
```

## random effect for sample of 50 pupils

```{r}
#rand effects
tidy(con_model, effects = "ran_vals", conf.int = TRUE) %>% sample_n(50) %>% 
ggplot(aes(level, estimate)) +
  geom_pointrange(aes(
    ymin = estimate - 1.96 * std.error,
    ymax = estimate + 1.96 * std.error
  )) + coord_flip() +
  theme_schola("x") +
  labs(title = "50 náhodně vybraných interceptů")
```

```{r}
con_model %>% ggpredict()
```

marginal plots draft

```{r}
# CI, fixed eff.
con_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl")) %>%
  plot + theme_schola("scatter")

con_model %>%
  ggpredict(terms = c("is_intervention2", "gender_girl", "age_fst_measur [meansd]")) %>% plot + theme_schola("scatter")

# one of the most interesting plots:
con_model %>%
  ggpredict(terms = c("age_fst_measur", "is_intervention2", "wave_ind")) %>%
  plot + theme_schola("scatter")

# predictiton interval (random effects)
con_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl"), type = "re") %>%
  plot + theme_schola("scatter")
```

# hyp


```{r}
hyp_model <- lmer(hyp ~ wave_ind * is_intervention2 + age_fst_measur + gender_girl + (wave_ind | id_pupil), data = df_long_zero)
summary(hyp_model)

df_long_zero %>% drop_na(hyp_model@frame %>% colnames) %>%
  mutate(model_residuals = resid(hyp_model), pred_vals = predict(hyp_model)) %>%
ggplot(aes(pred_vals, model_residuals)) +
  geom_point(alpha = .15) +
  geom_vline(xintercept = 0, col = "red") +
  geom_hline(yintercept = 0, col = "red") +
  theme_schola("scatter") +
  labs(title = "Reziduály fitovaného modelu")

```


```{r}
resid(hyp_model) %>%
  as_tibble %>%
  mutate(id = 1:n()) %>%
  ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  theme_schola("scatter") +
  labs(title = "Q-Q plot reziduálů")

# residuals densities
df_long_zero %>% drop_na(hyp_model@frame %>% colnames) %>%
  mutate(model_residuals = residuals(hyp_model)) %>%
ggplot(aes(x = model_residuals)) +
  geom_density(aes(color = is_intervention2), size = 1.25) +
  xlab("Residuals") +
  theme_schola("x") +
  labs(title = "Hustoty reziduálů dle skupiny")

# random effects of id_pupils to long format
ranef(hyp_model)$id_pupil %>%
  mutate(id = 1:n()) %>%
  pivot_longer(-id, "variable", "value") %>%
ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  facet_wrap(~ variable, scales = 'free_y') +
  theme_schola("scatter") +
  labs(title = "Q-Q ploty random efektů")

```

## Vizualizace modelu -- průměry 

```{r}
df_long_zero %>%
  drop_na(hyp_model@frame %>% colnames) %>%
  mutate(pred_val = predict(hyp_model, re.form = NA)) %>% # fixed eff only
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = hyp,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data - fixed effects")

df_long_zero %>%
  drop_na(hyp_model@frame %>% colnames) %>%
  mutate(pred_val = predict(hyp_model)) %>%
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = hyp,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data")
```

## random effect for sample of 50 pupils

```{r}
#rand effects
tidy(hyp_model, effects = "ran_vals", conf.int = TRUE) %>% sample_n(50) %>% 
ggplot(aes(level, estimate)) +
  geom_pointrange(aes(
    ymin = estimate - 1.96 * std.error,
    ymax = estimate + 1.96 * std.error
  )) + coord_flip() +
  theme_schola("x") +
  labs(title = "50 náhodně vybraných interceptů")
```

```{r}
hyp_model %>% ggpredict()
```

marginal plots draft

```{r}
# CI, fixed eff.
hyp_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl")) %>%
  plot + theme_schola("scatter")

hyp_model %>%
  ggpredict(terms = c("is_intervention2", "gender_girl", "age_fst_measur [meansd]")) %>% plot + theme_schola("scatter")

# one of the most interesting plots:
hyp_model %>%
  ggpredict(terms = c("age_fst_measur", "is_intervention2", "wave_ind")) %>%
  plot + theme_schola("scatter")

# predictiton interval (random effects)
hyp_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl"), type = "re") %>%
  plot + theme_schola("scatter")
```

# pee


```{r}
pee_model <- lmer(pee ~ wave_ind * is_intervention2 + age_fst_measur + gender_girl + (wave_ind | id_pupil), data = df_long_zero)
summary(pee_model)

df_long_zero %>% drop_na(pee_model@frame %>% colnames) %>%
  mutate(model_residuals = resid(pee_model), pred_vals = predict(pee_model)) %>%
ggplot(aes(pred_vals, model_residuals)) +
  geom_point(alpha = .15) +
  geom_vline(xintercept = 0, col = "red") +
  geom_hline(yintercept = 0, col = "red") +
  theme_schola("scatter") +
  labs(title = "Reziduály fitovaného modelu")

```


```{r}
resid(pee_model) %>%
  as_tibble %>%
  mutate(id = 1:n()) %>%
  ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  theme_schola("scatter") +
  labs(title = "Q-Q plot reziduálů")

# residuals densities
df_long_zero %>% drop_na(pee_model@frame %>% colnames) %>%
  mutate(model_residuals = residuals(pee_model)) %>%
ggplot(aes(x = model_residuals)) +
  geom_density(aes(color = is_intervention2), size = 1.25) +
  xlab("Residuals") +
  theme_schola("x") +
  labs(title = "Hustoty reziduálů dle skupiny")

# random effects of id_pupils to long format
ranef(pee_model)$id_pupil %>%
  mutate(id = 1:n()) %>%
  pivot_longer(-id, "variable", "value") %>%
ggplot(aes(sample = value)) +
  geom_qq() +
  geom_qq_line() +
  facet_wrap(~ variable, scales = 'free_y') +
  theme_schola("scatter") +
  labs(title = "Q-Q ploty random efektů")

```

## Vizualizace modelu -- průměry 

```{r}
df_long_zero %>%
  drop_na(pee_model@frame %>% colnames) %>%
  mutate(pred_val = predict(pee_model, re.form = NA)) %>% # fixed eff only
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = pee,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data - fixed effects")

df_long_zero %>%
  drop_na(pee_model@frame %>% colnames) %>%
  mutate(pred_val = predict(pee_model)) %>%
  group_by(wave_ind, is_intervention2) %>%
  summarise(pred_mean = mean(pred_val)) %>%
  ggplot(aes(x = wave_ind)) +
  geom_boxplot(aes(
    y = pee,
    x = as.factor(wave_ind),
    col = is_intervention2
  ), data = df_long_zero) +
  geom_line(aes(y = pred_mean, col = is_intervention2), size = 1.5) +
  theme_schola("y") +
  labs(title = "Model a naměřená data")
```

## random effect for sample of 50 pupils

```{r}
#rand effects
tidy(pee_model, effects = "ran_vals", conf.int = TRUE) %>% sample_n(50) %>% 
ggplot(aes(level, estimate)) +
  geom_pointrange(aes(
    ymin = estimate - 1.96 * std.error,
    ymax = estimate + 1.96 * std.error
  )) + coord_flip() +
  theme_schola("x") +
  labs(title = "50 náhodně vybraných interceptů")
```

```{r}
pee_model %>% ggpredict()
```

marginal plots draft

```{r}
# CI, fixed eff.
pee_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl")) %>%
  plot + theme_schola("scatter")

pee_model %>%
  ggpredict(terms = c("is_intervention2", "gender_girl", "age_fst_measur [meansd]")) %>% plot + theme_schola("scatter")

# one of the most interesting plots:
pee_model %>%
  ggpredict(terms = c("age_fst_measur", "is_intervention2", "wave_ind")) %>%
  plot + theme_schola("scatter")

# predictiton interval (random effects)
pee_model %>%
  ggpredict(terms = c("wave_ind", "is_intervention2", "gender_girl"), type = "re") %>%
  plot + theme_schola("scatter")
```
