==> 02_internal_report.Rmd <==
---
title: "Vnitřní report k SDQ"
subtitle: "základní charakteristiky s důrazem na shrnutí 7. vlny"
author:
- tým Schola Empirica
- "kontakt: Jaromír Mazák, mazak@scholaempirica.org"
date: '`r format(Sys.time(), "%d. %B %Y")`'
output: 
  reschola::schola_word:
    fig_caption: yes
    toc: no
    number_sections: false
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(include = T) 
options(scipen = 10) # force R to show full numbers, not scientific notation
options("OutDec" = ",") # Czech decimal mark

library(reschola)
library(tidyverse)
library(magrittr)
library(scales)
library(naniar)
source("shared.R")
set_reschola_ggplot_fonts() # make ggplot2 use Roboto fonts without you having to set it
library(here)
library(gganimate)
library(transformr)
library(corrr)
library(officer)
library(lme4)
library(lmerTest)
library(broom.mixed)
library(flextable)
library(ggeffects)

lmer <- lmerTest::lmer
filter <- dplyr::filter
here <- here::here
```

```{r}
# table numbering
tabs <- run_autonum("tab", "Tab. ")

# fig. numbering
fig_count = 0

fig_counter <- function(title = "default title") {
  fig_count <<- fig_count + 1
  return(paste0("Graf ", fig_count, ": ", title))
}
```

```{r}
# custom render engine to print tables "in loop"
render_custom <- function(x, ...) {
# the result of `knit_print.flextable`
  map(x, ~ knitr::knit_print(.x)) %>% 
    # collapse all the results
    paste(collapse = "\n") %>% 
    # knitr must use this result `asis`
    knitr::asis_output()
}





FitFlextableToPage <- function(ft, pgwidth = 6){

  ft_out <- ft %>% autofit()

  ft_out <- width(ft_out, width = dim(ft_out)$widths*pgwidth /(flextable_dim(ft_out)$widths))
  return(ft_out)
}
```

```{r}
# purr map generates ugly list indices when printing several objects, like [[1]]
# use suppress = TRUE directive in chunk header to suppress this
def <- knitr::knit_hooks$get("output")
knitr::knit_hooks$set(output = function(x, options) {
  x <- def(x, options)
  ifelse(!is.null(options$suppress), gsub(pattern = "```.*```", "", x), x)
})
```

```{r}
dim_labs <- c("emoce", "chování", "hyperaktivita", "vrstevníci", "prosociální")
names(dim_labs) = c("emo", "con", "hyp", "pee", "pro")

gender_labs <- c("dívky", "chlapci")
names(gender_labs) = c("girls", "boys")

intervention_labs <- c("kontrolní", "intervenční")
names(intervention_labs) = c("control", "experimental")

sel_age_labs <- paste0("vstupní věk = ", 3:7)
names(sel_age_labs) <- c("age = 3", "age = 4", "age = 5", "age = 6", "age = 7")
```

```{r data import}
dfm <- read_rds(here("data-processed/waves_1-7_cleaned-plus-IRT.rds"))
```

```{r}
read_csv(here("data-processed/cleaning_impact_wave.csv")) %>%
  flextable %>%
  colformat_num(j = "% lost", digits = 2, suffix = " %") %>%
  set_header_labels(
    wave = "glob. vlna",
    "# raw" = "poč. obs. před",
    "# cleaned" = "poč. obs. po",
    "# lost" = "poč. ztrac. obs.",
    "% lost" = "pod. ztrac. obs."
  ) %>% 
  padding(part = "footer", padding.bottom = 48) %>% 
  add_footer_lines("") %>%
  set_caption("Dopad validace dat - globální vlny", autonum = tabs) %>% 
  autofit

read_csv(here("data-processed/cleaning_impact_wave_ind.csv")) %>%
  flextable %>%
  colformat_num(j = "% lost", digits = 2, suffix = " %") %>%
  set_header_labels(
    wave_ind = "individ. vlna",
    "# raw" = "poč. obs. před",
    "# cleaned" = "poč. obs. po",
    "# lost" = "poč. ztrac. obs.",
    "% lost" = "pod. ztrac. obs."
  ) %>% padding(part = "footer", padding.bottom = 48) %>% 
  add_footer_lines("") %>% 
  set_caption("Dopad validace dat - individuální vlny", autonum = tabs) %>% 
  autofit
```


```{r include=FALSE}
# manual look
dfm$id_class %>% unique() %>% sort

# unite case
dfm %<>% mutate(id_class = str_to_lower(id_class))

# resolve "myšky" and "poupata"
dfm %<>% mutate(id_class = recode(id_class,
                                 "poupe_modra" = "modra_poupata",
                                 "poupe_zluta" = "zluta_poupata",
                                 "poupe_cervena" = "cervena_poupata",
                                 "myšky" = "mysky"
                                 ))

# replace "-" with proper NA...
dfm %<>% replace_with_na(replace = list(id_class = "-"))
```


```{r include=FALSE}
# from zscores to sten (0-10 scale), 5.5is the mean score of "normative" group (SDQ)
dfm %<>% mutate_at(vars(emo:pro), list(~ (. * 2) + 5))
```


```{r}
# global
dfm %>% group_by(wave) %>%
  summarise(
    children = n_distinct(id_pupil),
    n_interven = n_distinct(id_pupil[is_intervention2 == "experimental"]),
    sex = n_distinct(id_pupil[gender_girl == "girls"]),
    teachers  = n_distinct(id_teach_unq),
    classes = n_distinct(id_class),
    schools = n_distinct(id_school)
  ) %>% flextable() %>% set_header_labels(
    wave = "glob. vlna",
    children = "děti",
    n_interven = "v interv. skup.",
    sex = "dívky",
    teachers = "učitelé",
    classes = "třídy",
    schools = "školky"
  ) %>% 
  padding(part = "footer", padding.bottom = 48) %>% 
  add_footer_lines("") %>%
  set_caption("Základní popis - globální vlny", autonum = tabs) %>% 
  autofit()

# individual
dfm %>% group_by(wave_ind) %>%
  summarise(
    children = n_distinct(id_pupil),
    n_interven = n_distinct(id_pupil[is_intervention2 == "experimental"]),
    sex = n_distinct(id_pupil[gender_girl == "girls"]),
    teachers  = n_distinct(id_teach_unq),
    classes = n_distinct(id_class),
    schools = n_distinct(id_school)
  ) %>% flextable() %>% set_header_labels(
    wave_ind = "individ. vlna",
    children = "děti",
    n_interven = "v interv. skup.",
    sex = "dívky",
    teachers = "učitelé",
    classes = "třídy",
    schools = "školky"
  ) %>% 
  padding(part = "footer", padding.bottom = 48) %>% 
  add_footer_lines("") %>%
  set_caption("Základní popis - individuální vlny", autonum = tabs) %>% 
  autofit()

```



```{r}
dfm %>% group_by(wave) %>%
  summarise_at(vars(age, age_fst_measur),
               list(
                 mean = mean,
                 sd = sd,
                 min = min,
                 max = max
               ),
               na.rm = TRUE) %>%
  mutate(wave = as.character(wave)) %>%
  flextable() %>%
  colformat_num(digits = 2, na_str = "NA") %>%
  set_header_labels(
    wave = "glob. vlna",
    age_mean = "věk (prům.)",
    age_fst_measur_mean = "vstupní věk (prům.)",
    age_sd = "věk (SD)",
    age_fst_measur_sd = "vstupní věk (SD)",
    age_min = "věk (min)",
    age_fst_measur_min = "vstupní věk (min)",
    age_max = "věk (max)",
    age_fst_measur_max = "vstupní věk (max)"
  ) %>% 
  padding(part = "footer", padding.bottom = 48) %>% 
  add_footer_lines("") %>%
  set_caption("Základní popis (spojité proměnné) - globální vlny", autonum = tabs)

dfm %>% group_by(wave_ind) %>%
  summarise_at(vars(age, age_fst_measur),
               list(
                 mean = mean,
                 sd = sd,
                 min = min,
                 max = max
               ),
               na.rm = TRUE) %>%
  mutate(wave_ind = as.character(wave_ind)) %>%
  flextable() %>%
  colformat_num(digits = 2, na_str = "NA") %>%
  set_header_labels(
    wave_ind = "individ. vlna",
    age_mean = "věk (prům.)",
    age_fst_measur_mean = "vstupní věk (prům.)",
    age_sd = "věk (SD)",
    age_fst_measur_sd = "vstupní věk (SD)",
    age_min = "věk (min)",
    age_fst_measur_min = "vstupní věk (min)",
    age_max = "věk (max)",
    age_fst_measur_max = "vstupní věk (max)"
  ) %>%
  padding(part = "footer", padding.bottom = 48) %>% 
  add_footer_lines("") %>%
  set_caption("Základní popis (spojité proměnné) - individuální vlny", autonum = tabs)
```

```{r}
wave_labs <- paste0(seq(1, max(dfm$wave)), ". vlna")
names(wave_labs) = seq(1, max(dfm$wave))

wave_ind_labs <- paste0(seq(1, max(dfm$wave_ind)), ". vlna")
names(wave_ind_labs) = seq(1, max(dfm$wave_ind))
```


```{r}
dfm %>% count(gender = gender_girl, intervention = is_intervention2, wave = wave) %>%
  ggplot(aes(x = intervention, y = n, fill = gender)) +
  geom_bar(position = "fill", stat = "identity") +
  scale_y_percent_cz() +
  scale_x_discrete(labels = intervention_labs) +
  scale_fill_discrete(name = "pohlaví", labels = gender_labs) +
  facet_wrap(wave ~ ., nrow = 2, labeller = labeller(wave = wave_labs)) +
  labs(title = fig_counter("zastoupení pohlaví dle studijní skupiny"),
       subtitle = "členěno dle globálních vln") +
  theme_schola(legend.position = "bottom", multiplot = TRUE)

dfm %>% count(gender = gender_girl, intervention = is_intervention2, wave = wave_ind) %>%
  ggplot(aes(x = intervention, y = n, fill = gender)) +
  geom_bar(position = "fill", stat = "identity") +
  scale_y_percent_cz() +
  scale_x_discrete(labels = intervention_labs) +
  scale_fill_discrete(name = "pohlaví", labels = gender_labs) +
  facet_wrap(wave ~ ., nrow = 2, labeller = labeller(wave = wave_ind_labs)) +
  labs(title = fig_counter("zastoupení pohlaví dle studijní skupiny"),
       subtitle = "členěno dle individuálních vln") +
  theme_schola(legend.position = "bottom", multiplot = TRUE)
```
```{r}


# age by study group (waves 1-3 have absent age)
dfm %>% ggplot(aes(age, col = is_intervention2)) +
  geom_density() +
  facet_wrap(~ wave, nrow = 2, labeller = labeller(wave = wave_labs)) +
  scale_color_discrete(name = "studijní skupina", labels = intervention_labs) +
  labs(title = fig_counter("rozložení věku"),
       subtitle = "členěno dle globálních vln") +
  theme_schola(legend.position = c(.875, 0.25), multiplot = TRUE)

# age by study group (waves 1-3 have absent age)
dfm %>% ggplot(aes(age, col = is_intervention2)) +
  geom_density() +
  facet_wrap(~ wave_ind, nrow = 2, labeller = labeller(wave_ind = wave_ind_labs)) +
  scale_color_discrete(name = "studijní skupina", labels = intervention_labs) +
  labs(title = fig_counter("rozložení věku"),
       subtitle = "členěno dle individuálních vln") +
  theme_schola(legend.position = "bottom", multiplot = TRUE)

# first age by study group
dfm %>% ggplot(aes(age_fst_measur, col = is_intervention2)) +
  geom_density() +
  facet_wrap(~ wave, nrow = 2, labeller = labeller(wave = wave_labs)) +
  scale_color_discrete(name = "studijní skupina", labels = intervention_labs) +
  labs(title = fig_counter("rozložení vstupního věku"),
       subtitle = "členěno dle globálních vln") +
  theme_schola(legend.position = c(.875, 0.25), multiplot = TRUE)

# first age by study group
dfm %>% ggplot(aes(age_fst_measur, col = is_intervention2)) +
  geom_density() +
  facet_wrap(~ wave_ind, nrow = 2, labeller = labeller(wave_ind = wave_ind_labs)) +
  scale_color_discrete(name = "studijní skupina", labels = intervention_labs) +
  labs(title = fig_counter("rozložení vstupního věku"),
       subtitle = "členěno dle individuálních vln") +
  theme_schola(legend.position = "bottom", multiplot = TRUE)



# beware of missingness pattern!
dfm %>% count(missing_birthday = !is.na(born), intervention = is_intervention2, wave = wave) %>%
  ggplot(aes(x = intervention, y = n, fill = missing_birthday)) +
  geom_bar(position = "fill", stat = "identity") +
  scale_y_percent_cz() +
  scale_x_discrete(labels = intervention_labs) +
  scale_fill_discrete(name = "údaje o narození", labels = c("chybí", "ano")) +
  facet_wrap(wave ~ ., nrow = 2, labeller = labeller(wave = wave_labs)) +
  labs(title = fig_counter("chybějící údaje o narození"),
       subtitle = "členěno dle globálních vln") +
  theme_schola(legend.position = "bottom", multiplot = TRUE)

# beware of missingness pattern!
dfm %>% count(missing_birthday = !is.na(born), intervention = is_intervention2, wave = wave_ind) %>%
  ggplot(aes(x = intervention, y = n, fill = missing_birthday)) +
  geom_bar(position = "fill", stat = "identity") +
  scale_y_percent_cz() +
  scale_x_discrete(labels = intervention_labs) +
  scale_fill_discrete(name = "údaje o narození", labels = c("chybí", "ano")) +
  facet_wrap(wave ~ ., nrow = 2, labeller = labeller(wave = wave_ind_labs)) +
  labs(title = fig_counter("chybějící údaje o narození"),
       subtitle = "členěno dle individuálních vln") +
  theme_schola(legend.position = "bottom", multiplot = TRUE)
```


```{r}
ctt_dim_labs <- c("emoce", "chování", "hyperaktivita", "vrstevníci", "prosociální")
names(ctt_dim_labs) = c("temotion10", "tconduct10", "thyper10", "tpeer10", "tprosoc10")


# 7th wave
dfm %>% filter(wave == 7) %>% 
  drop_na(gender_girl) %>%
  select(id_pupil, gender_girl, is_intervention2, contains("10")) %>%
  pivot_longer(contains("10")) %>%
  ggplot(aes(value, fill = gender_girl)) +
  geom_bar(position = "dodge2") +
  scale_fill_discrete(name = "pohlaví", labels = gender_labs) +
  facet_wrap(~ name, labeller = labeller(name = ctt_dim_labs)) +
  theme_schola(legend.position = "bottom", multiplot = TRUE) +
  labs(title = fig_counter("rozložení skórů SDQ v globální sedmé vlně"))
```


```{r}

# IRT SDQ static, too many facets and lines IMHO
dfm %>%
  drop_na(gender_girl) %>%
  mutate(wave_ind = as.factor(wave_ind)) %>%
  select(id_pupil, gender_girl, is_intervention2, wave_ind, emo:pro) %>%
  pivot_longer(emo:pro) %>%
  ggplot(aes(value, col = wave_ind)) +
  geom_density() +
  facet_grid(gender_girl ~ name,
             labeller = labeller(name = dim_labs, gender_girl = gender_labs)) +
  theme_schola("x", multiplot = TRUE, legend.position = "bottom") +
  guides(colour = guide_legend(nrow = 1)) +
  labs(title = fig_counter("SDQ - odhady latentních skórů"),
       subtitle = "členěno dle jednotlivých dimenzí, pohlaví a individuálních vln měření",
       col = "individuální vlna")

```

```{r}
dfm %>%
  drop_na(gender_girl) %>%
  mutate(wave_ind = as.factor(wave_ind)) %>%
  select(id_pupil, gender_girl, is_intervention2, wave_ind, emo:pro) %>%
  pivot_longer(emo:pro) %>%
  ggplot(aes(value, col = wave_ind)) +
  geom_density() +
  facet_grid(is_intervention2 ~ name,
             labeller = labeller(name = dim_labs, is_intervention2 = intervention_labs)) +
  theme_schola("x", multiplot = TRUE, legend.position = "bottom") +
  guides(colour = guide_legend(nrow = 1)) +
  labs(title = fig_counter("SDQ - odhady latentních skórů"),
       subtitle = "členěno dle jednotlivých dimenzí, studijní skupiny a individuálních vln měření",
       col = "individuální vlna")
```



```{r suppress = TRUE}
# more "ergonomic" key variable names 
dfmod <- dfm %>%
  mutate(
    wave = factor(wave_ind),
    group = factor(is_intervention2, labels = intervention_labs),
    child = id_pupil,
    gender = factor(gender_girl, labels = gender_labs),
    age = age_fst_measur,
    class = id_class,
    school = id_school,
    teacher = id_teach_unq
  ) %>%
  select(wave,
         group,
         child,
         gender,
         age,
         class,
         school,
         teacher,
         emo:pro,
         ends_with("10")) %>%
  filter(wave %in% c(1:3))

# WARNING! because control group has only 3 possible measurements so far,
# interaction between wave and group cannot be estimated for the other waves
# however, it is possible to treat wave as numeric and get simple predictions
# of all remaining "levels"
# the other way is to throw the waves out and work with MM/YYYY times


# interactions between wave, group, gender and/or age was tested insignificant,
# we keep them as simple covariates

mod <-
  tibble(dim = names(dim_labs),
         model = map(dim, ~ lmer(
           paste(., "~ wave * group + age + gender + (1 | teacher / child)"),
           dfmod
         )))

# sign. interactions sorted by abs eff. size
map2_dfr(mod$model,
         mod$dim,
         ~ .x %>% tidy %>% add_column(.before = 1, dim = .y)) %>% 
  filter(p.value <= .05 & str_detect(term, ":")) %>% 
  arrange(desc(abs(estimate)))

map2(
  .x = mod$model,
  .y = 1:5,
  ~ ggpredict(.x, c("wave", "group", "gender")) %>%
    ggplot(
      aes(
        x,
        predicted,
        group = group,
        col = group,
        ymin = conf.low,
        ymax = conf.high
      )
    ) +
    geom_point(position = position_dodge(.2), size = 1.7) +
    geom_errorbar(
      position = position_dodge(.2),
      width = .2,
      size = .7
    ) +
    geom_line(size = .7) +
    facet_wrap( ~ facet) +
    theme_schola(
      multiplot = TRUE,
      legend.title = element_blank()
      ) +
    scale_y_continuous(limits = c(0, 10), breaks = seq(0, 10, 1)) +
    scale_x_discrete(labels = wave_ind_labs[1:3]) +
    labs(
      title = fig_counter("SDQ – vývoj v čase"),
      subtitle = paste("Dimenze –", dim_labs[.y]),
      caption = "se statisticky kontrolovaným vlivem věku"
    )
)
```


```{r suppress = TRUE}
# age
map2(
  .x = mod$model,
  .y = 1:5,
  ~ .x %>%
    ggpredict(terms = c("gender", "group", "age [3:7, by = 2]")) %>%
    plot(colors = c("#F8766D", "#00BFC4"),
         dodge = .15) +
    facet_wrap(~ facet, labeller = labeller(facet = sel_age_labs)) +
    scale_color_discrete(name = "skupina", labels = intervention_labs) +
    scale_y_continuous(limits = c(1, 10), breaks = seq(1, 10, 1)) +
    theme_schola(
      "y",
      multiplot = TRUE,
      legend.position = "bottom"
    ) +
    labs(
      title = fig_counter("SDQ – hodnoty při různých úrovních vtupního věku dle modelu"),
      subtitle = paste("Dimenze –", dim_labs[.y]))
)


# one of the most interesting plots:
map2(
  .x = mod$model,
  .y = 1:5,
  ~ .x %>%  ggpredict(terms = c("age", "group", "wave")) %>%
    plot(colors = c("#F8766D", "#00BFC4"),
         dodge = .15) +
    facet_wrap(~ facet, labeller = labeller(facet = wave_ind_labs[1:3])) +
    scale_y_continuous(limits = c(1, 10), breaks = seq(1, 10, 1)) +
    theme_schola("scatter", multiplot = TRUE, legend.position = "bottom", legend.title = element_blank()) +
    labs(
      title = fig_counter("SDQ – asociace vstupního věku a dimenze dle individuálních vln"),
      subtitle = paste("Dimenze –", dim_labs[.y]))
)


map2(
  .x = mod$model,
  .y = 1:5,
  ~ .x %>% ggpredict(terms = c("group", "gender", "wave")) %>% 
    ggplot(aes(
    x,
    predicted,
    col = group,
    ymin = conf.low,
    ymax = conf.high
  )) +
    facet_wrap(~ facet) +
  geom_point(position = position_dodge(.1), size = 1.7) +
  geom_errorbar(position = position_dodge(.1), width = .1, size = .7) +
  scale_y_continuous(limits = c(0, 10), breaks = seq(0, 10, 1)) +
  theme_schola("y", multiplot = TRUE, legend.title = element_blank()) +
    labs(
      title = fig_counter("SDQ – asociace vstupního věku a dimenze dle individuálních vln"),
      subtitle = paste("Dimenze –", dim_labs[.y]),
      caption = "se statisticky kontrolovaným vlivem věku")
)[1]

map2(
  .x = mod$model,
  .y = 1:5,
  ~ .x %>% ggpredict(terms = c("group", "gender")) %>% 
    ggplot(aes(
    x,
    predicted,
    col = group,
    ymin = conf.low,
    ymax = conf.high
  )) +
  geom_point(position = position_dodge(.1), size = 1.7) +
  geom_errorbar(position = position_dodge(.1), width = .1, size = .7) +
  scale_y_continuous(limits = c(0, 10), breaks = seq(0, 10, 1)) +
  theme_schola("y", multiplot = TRUE, legend.title = element_blank()) +
    labs(
      title = fig_counter("SDQ – asociace vstupního věku a dimenze dle individuálních vln"),
      subtitle = paste("Dimenze –", dim_labs[.y]),
      caption = "se statisticky kontrolovaným vlivem věku")
)[1]



```





```{r eval=FALSE, include=FALSE}
# export whole environment to make animated plots in separate script (for comp. time reason)
save(list = ls(all.names = TRUE), file = here("data-intermediate", "envir.RData"), envir = .GlobalEnv)
```
