==> 03_animated_plots.Rmd <==
---
title: "Animated plots"
author: "Jan Netík"
date: "6 6 2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

options(scipen = 10) # force R to show full numbers, not scientific notation
options("OutDec" = ",") # Czech decimal mark

library(reschola)
library(tidyverse)
library(magrittr)
library(scales)
library(naniar)
source("shared.R")
set_reschola_ggplot_fonts() # make ggplot2 use Roboto fonts without you having to set it
library(here)
library(gganimate)
library(transformr)
library(corrr)
library(officer)
library(lme4)
library(lmerTest)
library(broom.mixed)
library(flextable)
library(ggeffects)
library(conflicted)
conflict_prefer("lmer", "lmerTest")
conflict_prefer("filter", "dplyr")
```

```{r data import}
dfm <- read_rds(here("data-processed/waves_1-7_cleaned-plus-IRT.rds"))
```

```{r include=FALSE}
# from zscores to sten (0-10 scale), 5.5is the mean score of "normative" group (SDQ)
dfm %<>% mutate_at(vars(emo:pro), list(~ (. * 2) + 5))
```



```{r}
dim_labs <- c("emoce", "chování", "hyperaktivita", "vrstevníci", "prosociální")
names(dim_labs) = c("emo", "con", "hyp", "pee", "pro")

gender_labs <- c("dívky", "chlapci")
names(gender_labs) = c("girls", "boys")

intervention_labs <- c("kontrolní", "intervenční")
names(intervention_labs) = c("control", "experimental")

sel_age_labs <- paste0("vstupní věk = ", 3:7)
names(sel_age_labs) <- c(3:7)
```

```{r}
# load(here("data-intermediate/envir.RData"))
```

```{r eval=FALSE, include=FALSE}
  
# jawdropping plot in motion
sdq_dist_anim <- dfm %>% group_by(wave) %>% 
  drop_na(gender_girl) %>%
  select(id_pupil, gender_girl, is_intervention2, contains("10")) %>%
  pivot_longer(contains("10")) %>%
  ggplot(aes(value, fill = gender_girl)) +
  geom_bar(position = "dodge2") +
  scale_fill_discrete(name = "pohlaví", labels = gender_labs) +
  facet_wrap(~ name, labeller = labeller(name = ctt_dim_labs)) +
  theme_schola(multiplot = TRUE, legend.position = c(.825, 0.3), plot.tag.position = c(.825, 0.2), plot.tag = element_text(size = 9)) +
  guides(fill = guide_legend(ncol = 2)) +
  labs(tag = "Globální vlna č. {closest_state}") +
  transition_states(wave, transition_length = 3, state_length = 3, wrap = FALSE) +
   ease_aes('cubic-in-out') 

# sdq_dist_anim %>% animate()

anim_save("sdq_dist_anim_prez.gif", animation = sdq_dist_anim, path = here("data-processed"), height = 3.251349, width = 5.409448818897638, units = "in", res = 400)
```

```{r}
# IRT SDQ animated
sdq_irt_dist_anim <- dfm %>%
  drop_na(gender_girl) %>% group_by(wave_ind) %>%
  select(id_pupil, gender_girl, is_intervention2, emo:pro) %>%
  pivot_longer(emo:pro) %>%
  ggplot(aes(value, col = gender_girl)) +
  geom_density() +
  scale_color_discrete(name = "pohlaví", labels = gender_labs) +
  facet_wrap(~ name, nrow = 2, labeller = labeller(name = dim_labs)) +
  labs(title = fig_counter("SDQ - odhady latentních skórů v čase - individuální vlna {closest_state}"),
       subtitle = "členěno dle jednotlivých dimenzí a pohlaví") +
  theme_schola("x",
               multiplot = TRUE,
               legend.position = c(.825, 0.25)) +
     ease_aes('cubic-in-out') +
  transition_states(wave_ind)

anim_save("sdq_irt_dist_anim_prez.gif", animation = sdq_irt_dist_anim, path = here("data-processed"))
```

```{r}
dfm %>%
  drop_na(gender_girl) %>% 
  select(id_pupil, gender_girl, is_intervention2, emo:pro) %>%
  pivot_longer(emo:pro) %>%
  ggplot(aes(value, col = is_intervention2)) +
  geom_density(key_glyph = draw_key_smooth) +
  scale_color_discrete(name = "skupina", labels = intervention_labs) +
  facet_wrap(~ name, nrow = 2, labeller = labeller(name = dim_labs)) +
    scale_x_continuous(n.breaks = 7) +

  labs(x = "úroveň dimenze vyjádřená v **obvyklém měřítku SDQ** (tj. *0-10 bodů, 5 = průměr*)",
       y = "„relativní četnost“") +
  guides(color = guide_legend(keyheight = 1, title.vjust = -.3)) +
  coord_cartesian(xlim = c(-.3,10.3), expand = FALSE) +
  theme_schola("x",
               axis.title.x = ggtext::element_markdown(),
               axis.title.y = element_text(),
               multiplot = TRUE,
               legend.position = c(.825, 0.25))
```
